Implement PCSS

+++ Packages/TinyRenderPipeline/ShaderLibrary/DeclareOpaqueTexture.hlsl
@@ -6,7 +6,7 @@
 TEXTURE2D(_CameraOpaqueTexture);
 SAMPLER(sampler_CameraOpaqueTexture);
 
-float3 SampleSceneColor(float2 uv)
+half3 SampleSceneColor(float2 uv)
 {
     return SAMPLE_TEXTURE2D_LOD(_CameraOpaqueTexture, sampler_CameraOpaqueTexture, uv, 0.0).rgb;
 }
